Being overly paranoid about camelCase
Updating the name of the file to polyChains event though I believe the issue might be with people using dash, underscores, or any number of other mixed things. But in the off chance that possibly something else breaks up the names via this info etc I am renaming the file.